+++ src/index.js
@@ -4,16 +4,7 @@ import createHistory from 'history/createHashHistory';
 // import createHistory from 'history/createBrowserHistory';
 // 引入路由
 import router from './router';
-<<<<<<< HEAD
-// 引入model
-import home from './models/home';
-// 引入全局样式
-import './index.less';
-import './utils/axiosRequest';
-// 引入mock
-=======
 // 引入全局样式, mock, 拦截请求,设置token信息等
->>>>>>> master
 import './mock';
 import './utils/axiosRequest';
 import './index.less';