+++ src/router.js
@@ -1,15 +1,6 @@
 import React from 'react';
 import { Router, Route, Switch } from 'dva/router';
-<<<<<<< HEAD
-import loadable from 'utils/loadable';
-// import Authorized from 'components/common/Authorized';
-
-// 权限路由
-// const { AuthorizedRoute } = Authorized;
-
-=======
 import { loadable } from 'utils';
->>>>>>> master
 // 直接加载
 // import HomePage from './routes/HomePage';