+++ src/services/home.js
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 import axios from 'axios';
 
 export function getUser(params) {
@@ -7,15 +6,4 @@ export function getUser(params) {
 
 export function getList() {
   return axios.get('/api/list').catch(error => error);
-=======
-import { config } from 'utils';
-import axios from 'axios';
-
-export function getUser(params) {
-  return axios.get(`${config.host}/api/getUser`, params).catch(error => error);
-}
-
-export function getList() {
-  return axios.get(`${config.host}/api/list`, {isCancelTocken: true}).catch(() => ({error: true}));
->>>>>>> master
 }