+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - 新增模块热替换插件 `react-hot-loader`
   - 移除 `babel-plugin-dva-hmr` 插件，因为在 `webpack-server-dev` 下似乎并未起到任何有效作用
   - 针对 `webpack.config.js` 中潜在问题做了优化和修复
+  - 删除原来 `public/index.html` 并新增模板文件 `src/index.ejs`， 其他配置不变
   - 删除多线程压缩插件 `webpack-parallel-uglify-plugin`, 由于可能会在压缩后 `manifest.js` 的 `hash` 值不一致，导致页面访问报错
   - 多线程打包功能并不理想，删除 `happypack` ，原有的多线程 `happypack` 方案移动到新分支 `webpack-antd-happypack`
   - 使用 `uglifyjs-webpack-plugin` 替换原有的压缩方案， 压缩效果更好，且压缩时间有所缩减